Updating scritps & docs

+++ scripts/run_eval_queries.sh
@@ -1,5 +1,7 @@
 #/bin/bash
 YAHOO_ANSWERS_SOURCE="yahoo_answers"
+# Set this variable to 1, if you want to compute P@1 and recall@10 instead of ERR@10 and NDCG@20
+USE_OLD_STYLE_EVAL_FOR_YAHOO_ANSWERS="0"
 export MAVEN_OPTS="-Xms8192m -server"
 input=$1
 if [ "$input" = "" ] ; then
@@ -42,10 +44,15 @@ if [ "$DO_EVAL" = "" ] ; then
   exit 1
 fi
 
+YAHOO_STYLE_EVAL="0"
+if [ "$source_type" = "$YAHOO_ANSWERS_SOURCE" -a "$USE_OLD_STYLE_EVAL_FOR_YAHOO_ANSWERS" = "1" ] ; then
+  YAHOO_STYLE_EVAL="1"
+fi
+
 if [ "$DO_EVAL" = "1" -a "$source_type" != "$YAHOO_ANSWERS_SOURCE" ] ; then
   QREL_FILE_SHORT=$7
   if [ "$QREL_FILE_SHORT" = "" ] ; then
-    echo "The source type is different from $YAHOO_ANSWERS_SOURCE, hence, you have specify an external qrel file (7th argument)"
+    echo "The source type is different from $YAHOO_ANSWERS_SOURCE, hence, you have specify an external QREL file (7th argument)"
     exit 1
   fi
   if [ ! -f "$QREL_FILE_SHORT" ] ; then
@@ -57,38 +64,39 @@ fi
 # Retrieve 100 entries
 N=100
 
-TREC_EVAL_VER="9.0.4"
-
-TREC_EVAL_DIR="trec_eval-${TREC_EVAL_VER}"
-if [ ! -d $TREC_EVAL_DIR -o ! -f "$TREC_EVAL_DIR/trec_eval" ] ; then
-  rm -rf "$TREC_EVAL_DIR"
-  echo "Downloading and building missing trec_eval" 
-  wget https://github.com/usnistgov/trec_eval/archive/v${TREC_EVAL_VER}.tar.gz
-  if [ "$?" != "0" ] ; then
-    echo "Error downloading trec_eval"
-    exit 1
+if [ "$DO_EVAL" = "1" -a "$YAHOO_STYLE_EVAL" = "1" ] ; then
+  TREC_EVAL_VER="9.0.4"
+  TREC_EVAL_DIR="trec_eval-${TREC_EVAL_VER}"
+  if [ ! -d $TREC_EVAL_DIR -o ! -f "$TREC_EVAL_DIR/trec_eval" ] ; then
+    rm -rf "$TREC_EVAL_DIR"
+    echo "Downloading and building missing trec_eval" 
+    wget https://github.com/usnistgov/trec_eval/archive/v${TREC_EVAL_VER}.tar.gz
+    if [ "$?" != "0" ] ; then
+      echo "Error downloading trec_eval"
+      exit 1
+    fi
+    tar -zxvf v${TREC_EVAL_VER}.tar.gz
+    if [ "$?" != "0" ] ; then
+      echo "Error unpacking the trec_eval archive!"
+      exit 1
+    fi
+    cd $TREC_EVAL_DIR
+    if [ "$?" != "0" ] ; then
+      echo "Cannot changed dir to $TREC_EVAL_VER"
+      exit 1
+    fi
+    make
+    if [ "$?" != "0" ] ; then
+      echo "Error building trec_eval"
+      exit 1
+    fi
+    cd -
+    if [ "$?" != "0" ] ; then
+      echo "Cannot change dir back to the starting dir"
+      exit 1
+    fi
+    rm v${TREC_EVAL_VER}.tar.gz
   fi
-  tar -zxvf v${TREC_EVAL_VER}.tar.gz
-  if [ "$?" != "0" ] ; then
-    echo "Error unpacking the trec_eval archive!"
-    exit 1
-  fi
-  cd $TREC_EVAL_DIR
-  if [ "$?" != "0" ] ; then
-    echo "Cannot changed dir to $TREC_EVAL_VER"
-    exit 1
-  fi
-  make
-  if [ "$?" != "0" ] ; then
-    echo "Error building trec_eval"
-    exit 1
-  fi
-  cd -
-  if [ "$?" != "0" ] ; then
-    echo "Cannot change dir back to the starting dir"
-    exit 1
-  fi
-  rm v${TREC_EVAL_VER}.tar.gz
 fi
 
 for type in standard fixed ; do
@@ -97,8 +105,6 @@ for type in standard fixed ; do
     echo "There is no directory $INDEX_DIR"
     exit 1
   fi
-
-  RUN_DIR=
 
   if [ "$source_type" = "$YAHOO_ANSWERS_SOURCE" ] ; then
     QREL_FILE="$output/$type/runs/qrels.txt"
@@ -119,6 +125,7 @@ for type in standard fixed ; do
     flag=" -bm25fixed "
   fi
 
+  mkdir -p "$output/$type/runs"
   OUT_FILE="$output/$type/runs/trec_run"
   LOG_FILE="$output/$type/query.log"
   echo > $LOG_FILE
@@ -126,17 +133,23 @@ for type in standard fixed ; do
     echo "Error writing to $LOG_FILE"
     exit 1
   fi
-  for ((i=0;i<$REP_QTY;i++)) ; do
-    echo "Query iteration $(($i+1))"
-    scripts/lucene_query.sh -s data/stopwords.txt -i "$input" -source_type "$source_type" -d "$INDEX_DIR" -prob 1.0 -n $N -max_query_qty "$max_query_qty" -o "$OUT_FILE" $flag 2>&1 >> ${LOG_FILE}
+  if [ ! -f "$OUT_FILE" ] ; then
+    echo "Re-running Lucene"
+    for ((i=0;i<$REP_QTY;i++)) ; do
+      echo "Query iteration $(($i+1))"
+      scripts/lucene_query.sh -s data/stopwords.txt -i "$input" -source_type "$source_type" -d "$INDEX_DIR" -prob 1.0 -n $N -max_query_qty "$max_query_qty" -o "$OUT_FILE" $flag 2>&1 >> ${LOG_FILE}
 
-    if [ "$?" != "0" ] ; then
-      echo "lucene_query.sh failed!"
-      exit 1
-    fi
-  done
+      if [ "$?" != "0" ] ; then
+        echo "lucene_query.sh failed!"
+        exit 1
+      fi
+    done
+  else
+    echo "Re-using existing run: $OUT_FILE"
+  fi
   if [ "$DO_EVAL" = "1" ] ; then
     echo "Let's evaluate output quality"
+    EVAL_REPORT_PREFIX="$output/$type/runs/eval"
     if [ "$source_type" = "$YAHOO_ANSWERS_SOURCE" ] ; then
       # For Yahoo Answers's type of source queries
       # the qrel file will be huge, so we need to truncate it
@@ -147,18 +160,30 @@ for type in standard fixed ; do
         exit 1
       fi
     fi
-    EVAL_REPORT_PREFIX="$output/$type/runs/eval"
-    scripts/eval_output.py "$TREC_EVAL_DIR/trec_eval" "$QREL_FILE_SHORT" "$OUT_FILE" "$EVAL_REPORT_PREFIX"
-    if [ "$?" != "0" ] ; then
-      echo "eval_output failed!"
-      exit 1
+    if [ "$YAHOO_STYLE_EVAL" = "1" ] ; then
+      scripts/eval_output_trec_eval.py "$TREC_EVAL_DIR/trec_eval" "$QREL_FILE_SHORT" "$OUT_FILE" "$EVAL_REPORT_PREFIX"
+      if [ "$?" != "0" ] ; then
+        echo "scripts/eval_output_trec_eval.py failed!"
+        exit 1
+      fi
+    else
+      scripts/eval_output_gdeval.py "scripts/gdeval.pl" "$QREL_FILE_SHORT" "$OUT_FILE" "$EVAL_REPORT_PREFIX"
+      if [ "$?" != "0" ] ; then
+        echo "scripts/eval_output_gdeval.py failed!"
+        exit 1
+      fi
     fi
   fi
   
 done
 if [ "$DO_EVAL" = "1" ] ; then
   echo "Let's now compute p-values and ratios"
-  for metr in recall "recall@10" "P@1" map ; do
+  if [ "$YAHOO_STYLE_EVAL" = "1" ] ; then
+    metrics=(recall "recall@10" "P@1" map)
+  else
+    metrics=("ndcg@20" "err@20")
+  fi
+  for metr in ${metrics[*]} ; do
     echo "============================================="
     echo " Evaluation metric: $metr "
     echo "============================================="
